Split common crate by concern
